moved parameter to parameters.py and added test in test_parameters.py
